fix: frklstatus-start

+++ .dwm/frklstatus-start.sh
@@ -3,15 +3,21 @@
 PIDFILE="/tmp/frklstatus.pid"
 
 cleanup() {
-    rm -f "$PIDFILE"
+    if [ -f "$PIDFILE" ] && [ "$(cat "$PIDFILE")" = "$$" ]; then
+        rm -f "$PIDFILE"
+    fi
+    pgrep -f frklstatus
     exit 0
 }
 trap cleanup INT TERM EXIT
 #
 if [ -f "$PIDFILE" ]; then
+    echo "$PIDFILE pid arquivo"
     OLD_PID=$(cat "$PIDFILE")
+    echo "$OLD_PID pid antigo"
     if kill -0 "$OLD_PID" 2>/dev/null; then
-        exit 1
+        echo "$PIDFILE"
+        kill "$OLD_PID" 2>/dev/null
     fi
 fi
 
@@ -21,3 +27,4 @@ while true; do
     xsetroot -name "$(~/.dwm/frklstatus)"
     sleep 1s
 done
+
